Added specialties tab, refactored forms and fixed date issues with events tab
