+++ index-editor-app/index-editor-app-engine/JsonClasses/MembersPage.cs
@@ -13,7 +13,6 @@ namespace index_editor_app_engine
 {
     using System;
     using System.Collections.Generic;
-
     using System.Globalization;
     using Newtonsoft.Json;
     using Newtonsoft.Json.Converters;
