+++ index-editor-app/index-editor-app/index-editor-app.csproj.user
@@ -4,6 +4,9 @@
     <Compile Update="CreateSpecialtyForm.cs">
       <SubType>Form</SubType>
     </Compile>
+    <Compile Update="EventsForm.cs">
+      <SubType>Form</SubType>
+    </Compile>
     <Compile Update="Form1.cs">
       <SubType>Form</SubType>
     </Compile>
